Add response decryption functionality to GatewayFilters
A new parameter, `decryptResponse`, is added to the `ProxyRouteConfigDTO` class to check whether to decrypt responses or not. Adjustments were made in the `DecryptGatewayFilter` and `GatewayFiltersConfig` for this new functionality. An option to avoid decoding when `decryptResponse` is set to false was also added. The project version was updated to 15.4-RELEASE to reflect these changes.