Update exception handling in RequestHistory services
Reworked the exception handling in RequestHistoryExceptionHandler and RequestHistoryDBService. Previously, errors could occur when a request history was not set, resulting in potential application crashes. Now, a new CompletableFuture is created if no request history exists. Additionally, made changes to the saving of authentication information to ensure it properly handles cases where request history is not set.